+++ tests/test_utils.py
@@ -1,11 +1,16 @@
 import asyncio
+import json
 import logging
+from pathlib import Path
 
 import pytest
 from _pytest.capture import CaptureFixture
 from _pytest.logging import LogCaptureFixture
 
-from ytdl_api.utils import repeat_at
+from ytdl_api.utils import (
+    get_po_token_verifier_from_file,
+    repeat_at,
+)
 
 
 @pytest.fixture
@@ -48,3 +53,22 @@ async def test_repeat_at_with_logger(caplog: LogCaptureFixture, logger: logging.
     assert len(captured_logs) > 0
     assert hasattr(captured_logs[0], "exc_text")
     assert 'raise Exception("Hello")' in captured_logs[0].exc_text
+
+
+def test_get_po_token_verifier():
+    """
+    Test Case for get_po_token_verifier
+    """
+    valid_po_token_file = Path(__file__).parent / "fixtures" / "valid_po_token.json"
+    parsed_valid_po_token_file: dict[str, str] = json.loads(valid_po_token_file.read_text("utf-8"))
+    visitor_data, po_token = get_po_token_verifier_from_file(valid_po_token_file)
+    assert visitor_data == parsed_valid_po_token_file["visitorData"]
+    assert po_token == parsed_valid_po_token_file["poToken"]
+
+
+def test_get_po_token_verifier_invalid_file():
+    """
+    Test Case for get_po_token_verifier with invalid file
+    """
+    with pytest.raises(Exception):
+        get_po_token_verifier_from_file(Path(__file__).parent / "fixtures" / "invalid_po_token.json")