+++ ytdl_api/config.py
@@ -19,6 +19,8 @@ from .utils import LOGGER
 REPO_PATH = (Path(__file__).parent / "..").resolve()
 ENV_PATH = (REPO_PATH / ".env").resolve()
 
+PO_TOKEN_FILE_PATH = Path(__file__).parent / "po_token_verifier.json"
+
 
 class DetaBaseDataSourceConfig(BaseConfig):
     """
@@ -91,6 +93,10 @@ class Settings(BaseConfig):
 
     @property
     def downloader_version(self) -> str:
+        if self.downloader == DownloaderType.PYTUBE:
+            return version("pytubefix")
+        if self.downloader == DownloaderType.MOCK:
+            return "mock"
         return version(self.downloader.value)
 
     @validator("remove_expired_downloads_task_cron", pre=True)