+++ ytdl_api/dependencies.py
@@ -13,9 +13,9 @@ from .callbacks import (
     on_start_converting,
     on_ytdlp_progress_callback,
 )
-from .config import Settings
+from .config import PO_TOKEN_FILE_PATH, Settings
 from .constants import DownloaderType
-from .utils import LOGGER
+from .utils import LOGGER, get_po_token_verifier_from_file
 
 
 # Ignoring get_settings dependency in coverage because it will be
@@ -66,14 +66,11 @@ def get_ytdlp_downloader(
     )
 
 
-def get_downloader(
-    settings: Settings = Depends(get_settings),
-    datasource: datasource.IDataSource = Depends(get_database),
-    event_queue: queue.NotificationQueue = Depends(get_notification_queue),
-    storage: storage.IStorage = Depends(get_storage),
-) -> downloaders.IDownloader:
-    if settings.downloader == DownloaderType.YTDLP:
-        return get_ytdlp_downloader(datasource, event_queue, storage)
+def get_pytube_downloader(
+    datasource: datasource.IDataSource,
+    event_queue: queue.NotificationQueue,
+    storage: storage.IStorage,
+):
     on_download_started_hook = partial(on_download_start_callback, datasource=datasource, queue=event_queue)
     on_progress_hook = partial(on_pytube_progress_callback, datasource=datasource, queue=event_queue)
     on_converting_hook = partial(on_start_converting, datasource=datasource, queue=event_queue)
@@ -96,7 +93,20 @@ def get_downloader(
         on_converting_callback=on_converting_hook,
         on_finish_callback=on_finish_hook,
         on_error_callback=on_error_hook,
+        po_token_verifier=partial(get_po_token_verifier_from_file, PO_TOKEN_FILE_PATH),
+        logger=LOGGER,
     )
+
+
+def get_downloader(
+    settings: Settings = Depends(get_settings),
+    datasource: datasource.IDataSource = Depends(get_database),
+    event_queue: queue.NotificationQueue = Depends(get_notification_queue),
+    storage: storage.IStorage = Depends(get_storage),
+) -> downloaders.IDownloader:
+    if settings.downloader == DownloaderType.YTDLP:
+        return get_ytdlp_downloader(datasource, event_queue, storage)
+    return get_pytube_downloader(datasource, event_queue, storage)
 
 
 def get_uid_dependency_factory(raise_error_on_empty: bool = False):