+++ ytdl_api/exceptions.py
@@ -3,7 +3,7 @@ from http.client import RemoteDisconnected
 from logging import Logger
 
 from fastapi.requests import Request
-from pytube.exceptions import AgeRestrictedError, RegexMatchError, VideoPrivate
+from pytubefix.exceptions import AgeRestrictedError, LoginRequired, RegexMatchError, VideoPrivate
 from starlette.responses import JSONResponse
 from starlette.status import HTTP_401_UNAUTHORIZED, HTTP_403_FORBIDDEN, HTTP_500_INTERNAL_SERVER_ERROR
 
@@ -69,6 +69,15 @@ async def on_pytube_videoprivate_error(logger: Logger, request: Request, exc: Vi
     )
 
 
+async def on_login_required_error(logger: Logger, request: Request, exc: LoginRequired):
+    logger.exception(exc)
+    return make_internal_error(
+        "login-required",
+        exc.error_string,
+        HTTP_403_FORBIDDEN,
+    )
+
+
 ERROR_HANDLERS = (
     (RemoteDisconnected, on_remote_disconnected),
     (socket.timeout, on_socket_timeout),
@@ -76,5 +85,6 @@ ERROR_HANDLERS = (
     (AgeRestrictedError, on_pytube_agerestricted_error),
     (RegexMatchError, on_pytube_regexmatch_error),
     (VideoPrivate, on_pytube_videoprivate_error),
+    (LoginRequired, on_login_required_error),
     (Exception, on_default_exception_handler),
 )